Update README.md
# SLM_group_project
A group project. The purpose of this project is to find errors and improve the excisting matlab codes.
All the codes here are currently (20230305) work with the green laser (532nm)
For any other wavelength the user needs to make a few additional changes - calculate and save new holograms for the calibration code - circle, grid and a vortex.
Currently (20230305), to use the code, the user needs to proveid "SLM_Screen" before runnig the code. Can be found using gds function on matlab.
Currently (20230305), to use the code, all the information (calibration files and precalculated holograms) must be saved on the user computer and the user needs to provide the correct path in order to run the code.  
Tasks:
     1. We need to improve the code and maybe add a line that identifies the device number.
     2. We need to save the calibration files and other information on a cloud and add the correct path to the codes or maybe find a better way to do it.
     
     These are the problems I remember we talked about in our last meeting. 
     Fill free to add any suggestions or comments here. 
     Ron.